+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+
+<a name="v3.0.3"></a>
+## [v3.0.3] - 2022-01-08
 ### Feat
 - **alpha:** fortune now only get if exist
 - **alpha:** added Change Directory entry
@@ -478,7 +481,8 @@
 - **indentation:** moved indentation to it's own folder
 
 
-[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v3.0.2...HEAD
+[Unreleased]: https://github.com/tigorlazuardi/nvim/compare/v3.0.3...HEAD
+[v3.0.3]: https://github.com/tigorlazuardi/nvim/compare/v3.0.2...v3.0.3
 [v3.0.2]: https://github.com/tigorlazuardi/nvim/compare/v3.0.1...v3.0.2
 [v3.0.1]: https://github.com/tigorlazuardi/nvim/compare/v3.0.0...v3.0.1
 [v3.0.0]: https://github.com/tigorlazuardi/nvim/compare/v2.6.0...v3.0.0